purge and rewrite helpers

+++ src/npr/ont.py
@@ -1,4 +1,3 @@
-import glob
 import os
 import signal
 import sys
@@ -7,12 +6,11 @@ from pathlib import Path
 import subprocess as sp
 from threading import Event
 import rich_click as click
-import snakemake
 import yaml
 from rich import print
 from npr.communication import query_parkour, send_email, ship_qcreports, standard_text
 from npr.ont_pipeline import find_new_flowcell, get_periphery, read_flowcell_info, read_samplesheet
-from npr.snakehelper import getfast5foot, monitor_storage, scan_multiqc, config_to_smkcmd
+from npr.snakehelper import getfast5foot, get_disk_stat, scan_multiqc, config_to_smkcmd, print_header
 
 
 # set up CLI args.
@@ -150,25 +148,26 @@ def main(config):
 
         flowcell, msg, base_path = find_new_flowcell(config)
         if flowcell:
-            if (
-                "organism" not in config["info_dict"]
-                or "protocol" not in config["info_dict"]
-            ):
-                # need parkour query only if 'organism' or 'protocol' is undefined
-                msg = query_parkour(config, flowcell, msg)
+            print_header(flowcell)
+            print_header('parkour')
+            msg = query_parkour(config, flowcell, msg)
             # The following should be simplified but I did not want to touch
             # read_flow_cell_info() for nowq
+            print_header('read_flowcell_info')
             config["info_dict"] = read_flowcell_info(
                 config, config["info_dict"], base_path
             )
 
             # read samplesheet
+            print_header('read_samplesheet')
             bc_kit, data = read_samplesheet(config)
             config["data"] = data
             config["bc_kit"] = bc_kit
             print(config["data"])
             print("samplesheet is read sucessfully")
 
+            # Snakemake setup.
+            print_header('Starting snakemake')
             config["info_dict"]["logfile"] = os.path.join(
                 config["info_dict"]["flowcell_path"], "log", "ont.log"
             )
@@ -180,8 +179,6 @@ def main(config):
             config["info_dict"]["configFile"] = configFile
             with open(configFile, "w") as f:
                 yaml.dump(config, f, default_flow_style=False)
-
-            send_email("Found flowcell:", msg, config, allreceivers=False)
 
             config["info_dict"]["transfer_path"] = get_periphery(config)
 
@@ -201,40 +198,62 @@ def main(config):
                 ]
             )
             _smk_cmd.insert(0, 'snakemake')
-            print(_smk_cmd)
-            rg = sp.Popen(_smk_cmd, stdout=sp.PIPE, stderr=sp.PIPE, text=True)
-            stdout, stderr = rg.communicate()  # Wait for process to complete
-            if rg.returncode != 0:
-                msg += f"Snakemake errorcode {rg.returncode}\n"
-                msg += stderr
-                send_email("Snakemake failed for flowcell:", msg, config)
+            
+            #
+            print(f"Starting snakemake with command:\n {' '.join(_smk_cmd)}\n")
+            rg = sp.Popen(_smk_cmd, stdout=sp.PIPE, stderr=sp.STDOUT, text=True, bufsize=1, universal_newlines=True)
+            ols = []
+            for line in iter(rg.stdout.readline, ''):
+                print(line, end='')
+                ols.append(line)
+            rg.stdout.close()
+            smk_log = ''.join(ols)
+            return_code = rg.wait()
+
+            #stdout, stderr = rg.communicate()
+            if return_code != 0:
+                msg += f"Snakemake errorcode {return_code}\n"
+                msg += smk_log
+                send_email("Snakemake failed for flowcell:", msg, config, failure=True)
                 sys.exit(1)
+            else:
+                with open(os.path.join(config["info_dict"]["flowcell_path"], "snakemake.log"), 'w') as f:
+                    f.write(smk_log)
+            print(config)
+            print_header("Post snakemake processing")
 
-            msg += "pod5 compression: {}\n".format(
-                getfast5foot(
-                    config["info_dict"]["base_path"],
-                    config["info_dict"]["flowcell_path"],
+            if return_code == 0 and not config["snakemake"]["dryrun"]:
+                msg += "pod5 compression: {}\n".format(
+                    getfast5foot(
+                        config["info_dict"]["base_path"],
+                        config["info_dict"]["flowcell_path"],
+                    )
                 )
-            )
-            # spread the news
-            ship_qcreports(config, flowcell)
-            config["QC"] = scan_multiqc(config)
-            config["SM"] = monitor_storage(config)
-            msg = standard_text(config)
-            send_email("Successfully finished flowcell:", msg, config)
+                # spread the news
+                ship_qcreports(config, flowcell)
+                config["QC"] = scan_multiqc(config)
+                config["SM"] = {
+                                    "offload (deepseq)": get_disk_stat(config['paths']['offloadDir']),
+                                    "inbox (bioinfo)": get_disk_stat(config['paths']['outputDir']),
+                                    "periphery": get_disk_stat(config['info_dict']['transfer_path'], True)
+                                }
+            
+                msg = standard_text(config)
+                send_email("Successfully finished flowcell:", msg, config, failure=False)
 
-            summaryFile = os.path.join(
-                config["info_dict"]["flowcell_path"], "log/summary.yaml"
-            )
-            with open(summaryFile, "w") as f:
-                yaml.dump(config, f, default_flow_style=False)
+                summaryFile = os.path.join(
+                    config["info_dict"]["flowcell_path"], "log/summary.yaml"
+                )
+                with open(summaryFile, "w") as f:
+                    yaml.dump(config, f, default_flow_style=False)
 
-            Path(
-                os.path.join(config["info_dict"]["flowcell_path"], "analysis.done")
-            ).touch()
+                Path(
+                    os.path.join(config["info_dict"]["flowcell_path"], "analysis.done")
+                ).touch()
 
-            # wipe config['info_dict']
-            config["info_dict"] = {}
+                # wipe config['info_dict']
+                config["info_dict"] = {}
+                
         else:
             print("No flowcells found. I go back to sleep.")
             sleep()
